WP1. Build fetcher and decoder finish at 2/29/2024
top.sv to include a sequential instruction fetcher and decoder for the RV64IM instructions.  Your implementation should read and decode memory starting at the "entry" point address.  Use $display to output a line of text for each decoded instruction.  Each line should comprise the address of the instruction in hex, the decoded instruction mnemonic, and the number of the rd register if the instruction has one (<instruction address in hex>: <mnemonic> <rd>).  Use $finish; to terminate the simulation when an all-zero (64'b0) response is received from memory.

Test cases to pass can be found in /test/tests/wp1