WP 2 – Build an ALU, Finish at 3/22/2024
Design an ALU which performs the actual computation for the arithmetic instructions. Start with your code for wp1. You will have to add a Register File to maintain the state of the system. Don’t worry about loads and stores or branch instructions (jump, call, ret, etc.) in this warmup project. In the end, print out the 32 64-bit hexadecimal values, one per line, corresponding to the final values in the registers.

Test cases to pass can be found in /test/tests/wp2