5+ stage,Set-Associative Caches,4/4/2024
Unlike the warmup projects, we actually want the ability to perform IO operations. However, because we do not have an OS, we have to fake its existence. To do this, a Verilog function do_ecall() is provided. When your core encounters an ECALL instruction in the commit/writeback stage, perform the following call:

do_ecall(a7, a0, a1, a2, a3, a4, a5, a6, a0);
where all but the last a# arguments are the logic[63:0] objects holding the values of the corresponding registers, and the last a# argument is the object into which the return value will be written by the system call. After calling do_ecall, you must flush the pipeline, because this operation will set the last a0 argument to the return value of the system call, which requires all dependent instructions to be re-executed.

Additionally, to correctly support simulation of fake OS system calls that read memory contents in the presence of write queues and write-back caches, you must add another “magic” call to your system at commit/writeback to be performed on all store instructions:

do_pending_write(address, value, size);
where the size specifies the number of bytes being written. If using HAVETLB=y, the address should be physical.

Moreover, to correctly support system calls that write memory contents, you must add support to your data caches for invalidation requests. Specifically, your cache may receive transactions with MakeInvalid (0xD) on the acsnoop lines.  If your cache contains data lines covering the invalidated 64-byte range, you must mark these lines as invalid. Importantly (in a slight departure from how invalidations work in real processors), MakeInvalid transactions take precedence over any demand requests from the core and your cache is not permitted to accept read requests from the core when a MakeInvalid transaction is indicated.  NOTE: The invalidation mechanism is required in modern systems for I/O devices and for multi-core systems; in this project, we are abusing this mechanism to enable our fake OS simulation, syscalls that write data on real systems do not trigger invalidations of this type.

Test cases to pass can be found in /test/tests/project