Final code simplification
More explanation of ‘ecall’:

On a real system, an ecall will switch the CPU's "mode" from U-ser to S-upervisor and jump to a handler, which will execute a bunch of code in the operating system.  That code will perform whatever actions are needed and return from the ecall by switching the mode back to U-ser and continuing execution with the PC after the ecall instruction.

If the system call is an operation that writes to memory (e.g., a read() syscall that reads from a file and writes to the application memory), then quite a bit of code would run in the OS before returning to your application.  For example, a disk read using DMA would be requested, the application process put into a sleeping/waiting state, a context switch would occur and other programs could run on the core (or the OS idle loop would run for a while).  In the meantime, the DMA controller would be reading data from the disk and writing it into memory, where each write by the DMA controller would be issuing "invalidate" requests to the memory system to make sure that none of the caches in the system have a copy of the blocks that are being overwritten by the DMA.  This is needed to makes sure that subsequent Load instructions from the CPU experience cache misses, forcing them to read the data from memory (where the DMA controller wrote them).  When the DMA transfer is done, the DMA controller raises an interrupt to make the OS re-run the scheduler and wake up the sleeping process.  When the process resumes, it will run Load instructions, which will get cache misses and read data from memory (which the DMA controller wrote to).

In the course infrastructure, there is no OS.  Instead, you are asked to call a "magic" function do_ecall(), where the simulation infrastructure fakes the behavior of the requested system call.  When the system call is being emulated, if that system call writes to memory, then requests to invalidate the blocks being written are sent to your core.